🐛 FIX: Minor fixes
