+++ src/components/login/index.jsx
@@ -1,3 +1,2 @@
 export { default as LoginHeader } from 'components/login/LoginHeader'
 export { default as LoginBody } from 'components/login/LoginBody'
-export { default as LoginUtils } from 'components/login/LoginUtils'
